refactor(whats-new): align module boundaries
Keep release projection logic framework-free and colocate HTTP transport with its tests. Invalid checked-in metadata remains fail-open.\n\nRefs PUL-186
