+++ backend-nest/src/modules/whats-new/application/get-ios-whats-new.use-case.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import type { WhatsNewQuery, WhatsNewResponse } from 'pulpe-shared';
-import { buildWhatsNewResponse } from '../whats-new-payload';
+import { buildWhatsNewResponse } from '../domain/whats-new-payload';
 
 @Injectable()
 export class GetIosWhatsNewUseCase {
