+++ backend-nest/src/modules/whats-new/infrastructure/http/dto/whats-new-swagger.dto.ts
