+++ backend-nest/src/modules/whats-new/whats-new.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { GetIosWhatsNewUseCase } from './application/get-ios-whats-new.use-case';
-import { WhatsNewController } from './whats-new.controller';
+import { WhatsNewController } from './infrastructure/http/whats-new.controller';
 
 @Module({
   controllers: [WhatsNewController],
